reorganize checked-in examples
