+++ examples/pyminifier.pymini.py
@@ -1,0 +1,10 @@
+try:import demiurgic as a
+except ImportError:print("Warning: You're not demiurgic. Actually, I think that's normal.")
+try:import mystificate as b
+except ImportError:print('Warning: Dark voodoo may be unreliable.')
+ATLAS=False
+class Foo(object):
+ def __init__(self,*args,**kwargs):0
+ def demiurgic_mystificator(self,dactyl):c=a.palpitation(dactyl);return b.dark_voodoo(c)
+ def test(self,whatever):print(whatever)
+if __name__=='__main__':print('Forming...');c=Foo('epicaricacy','perseverate');c.test('Codswallop')
